+++ internal/kibana/synthetics/schema.go
@@ -11,6 +11,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/boolplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/int64planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
@@ -62,22 +64,37 @@ type tfTCPMonitorFieldsV0 struct {
 	ProxyUseLocalResolver types.Bool     `tfsdk:"proxy_use_local_resolver"`
 }
 
+type tfICMPMonitorFieldsV0 struct {
+	Host types.String `tfsdk:"host"`
+	Wait types.Int64  `tfsdk:"wait"`
+}
+
+type tfBrowserMonitorFieldsV0 struct {
+	InlineScript      types.String         `tfsdk:"inline_script"`
+	Screenshots       types.String         `tfsdk:"screenshots"`
+	SyntheticsArgs    []types.String       `tfsdk:"synthetics_args"`
+	IgnoreHttpsErrors types.Bool           `tfsdk:"ignore_https_errors"`
+	PlaywrightOptions jsontypes.Normalized `tfsdk:"playwright_options"`
+}
+
 type tfModelV0 struct {
-	ID               types.String           `tfsdk:"id"`
-	Name             types.String           `tfsdk:"name"`
-	SpaceID          types.String           `tfsdk:"space_id"`
-	Schedule         types.Int64            `tfsdk:"schedule"`
-	Locations        []types.String         `tfsdk:"locations"`
-	PrivateLocations []types.String         `tfsdk:"private_locations"`
-	Enabled          types.Bool             `tfsdk:"enabled"`
-	Tags             []types.String         `tfsdk:"tags"`
-	Alert            *tfAlertConfigV0       `tfsdk:"alert"`
-	APMServiceName   types.String           `tfsdk:"service_name"`
-	TimeoutSeconds   types.Int64            `tfsdk:"timeout"`
-	HTTP             *tfHTTPMonitorFieldsV0 `tfsdk:"http"`
-	TCP              *tfTCPMonitorFieldsV0  `tfsdk:"tcp"`
-	Params           jsontypes.Normalized   `tfsdk:"params"`
-	RetestOnFailure  types.Bool             `tfsdk:"retest_on_failure"`
+	ID               types.String              `tfsdk:"id"`
+	Name             types.String              `tfsdk:"name"`
+	SpaceID          types.String              `tfsdk:"space_id"`
+	Schedule         types.Int64               `tfsdk:"schedule"`
+	Locations        []types.String            `tfsdk:"locations"`
+	PrivateLocations []types.String            `tfsdk:"private_locations"`
+	Enabled          types.Bool                `tfsdk:"enabled"`
+	Tags             []types.String            `tfsdk:"tags"`
+	Alert            *tfAlertConfigV0          `tfsdk:"alert"`
+	APMServiceName   types.String              `tfsdk:"service_name"`
+	TimeoutSeconds   types.Int64               `tfsdk:"timeout"`
+	HTTP             *tfHTTPMonitorFieldsV0    `tfsdk:"http"`
+	TCP              *tfTCPMonitorFieldsV0     `tfsdk:"tcp"`
+	ICMP             *tfICMPMonitorFieldsV0    `tfsdk:"icmp"`
+	Browser          *tfBrowserMonitorFieldsV0 `tfsdk:"browser"`
+	Params           jsontypes.Normalized      `tfsdk:"params"`
+	RetestOnFailure  types.Bool                `tfsdk:"retest_on_failure"`
 }
 
 func GetCompositeId(id string) (*clients.CompositeId, diag.Diagnostics) {
@@ -165,12 +182,69 @@ func monitorConfigSchema() schema.Schema {
 				Optional:            true,
 				MarkdownDescription: "The monitor timeout in seconds, monitor will fail if it doesn’t complete within this time. Default: `16`",
 			},
-			"params": jsonObjectSchema("Monitor parameters"),
-			"http":   httpMonitorFieldsSchema(),
-			"tcp":    tcpMonitorFieldsSchema(),
+			"params":  jsonObjectSchema("Monitor parameters"),
+			"http":    httpMonitorFieldsSchema(),
+			"tcp":     tcpMonitorFieldsSchema(),
+			"icmp":    icmpMonitorFieldsSchema(),
+			"browser": browserMonitorFieldsSchema(),
 			"retest_on_failure": schema.BoolAttribute{
 				Optional:            true,
 				MarkdownDescription: "Enable or disable retesting when a monitor fails. By default, monitors are automatically retested if the monitor goes from \"up\" to \"down\". If the result of the retest is also \"down\", an error will be created, and if configured, an alert sent. Then the monitor will resume running according to the defined schedule. Using retest_on_failure can reduce noise related to transient problems. Default: `true`.",
+			},
+		},
+	}
+}
+
+func browserMonitorFieldsSchema() schema.Attribute {
+	return schema.SingleNestedAttribute{
+		Optional:            true,
+		MarkdownDescription: "Browser Monitor specific fields",
+		Attributes: map[string]schema.Attribute{
+			"inline_script": schema.StringAttribute{
+				Optional:            false,
+				Required:            true,
+				MarkdownDescription: "The inline script.",
+			},
+			"screenshots": schema.StringAttribute{
+				Optional:            true,
+				MarkdownDescription: "Controls the behavior of the screenshots feature.",
+				Validators: []validator.String{
+					stringvalidator.OneOf("on", "off", "only-on-failure"),
+				},
+				Computed:      true,
+				PlanModifiers: []planmodifier.String{stringplanmodifier.UseStateForUnknown()},
+			},
+			"synthetics_args": schema.ListAttribute{
+				ElementType:         types.StringType,
+				Optional:            true,
+				MarkdownDescription: "Synthetics agent CLI arguments.",
+			},
+			"ignore_https_errors": schema.BoolAttribute{
+				Optional:            true,
+				MarkdownDescription: "Whether to ignore HTTPS errors.",
+				Computed:            true,
+				PlanModifiers:       []planmodifier.Bool{boolplanmodifier.UseStateForUnknown()},
+			},
+			"playwright_options": jsonObjectSchema("Playwright options."),
+		},
+	}
+}
+
+func icmpMonitorFieldsSchema() schema.Attribute {
+	return schema.SingleNestedAttribute{
+		Optional:            true,
+		MarkdownDescription: "ICMP Monitor specific fields",
+		Attributes: map[string]schema.Attribute{
+			"host": schema.StringAttribute{
+				Optional:            false,
+				Required:            true,
+				MarkdownDescription: "Host to ping; it can be an IP address or a hostname.",
+			},
+			"wait": schema.Int64Attribute{
+				Optional:            true,
+				MarkdownDescription: " Wait time in seconds. Default: `1`",
+				PlanModifiers:       []planmodifier.Int64{int64planmodifier.UseStateForUnknown()},
+				Computed:            true,
 			},
 		},
 	}
@@ -228,6 +302,8 @@ func httpMonitorFieldsSchema() schema.Attribute {
 			"max_redirects": schema.Int64Attribute{
 				Optional:            true,
 				MarkdownDescription: "The maximum number of redirects to follow. Default: `0`",
+				PlanModifiers:       []planmodifier.Int64{int64planmodifier.UseStateForUnknown()},
+				Computed:            true,
 			},
 			"mode": schema.StringAttribute{
 				Optional:            true,
@@ -414,6 +490,8 @@ func (v *tfModelV0) toModelV0(api *kbapi.SyntheticsMonitor) (*tfModelV0, error) 
 
 	var http *tfHTTPMonitorFieldsV0
 	var tcp *tfTCPMonitorFieldsV0
+	var icmp *tfICMPMonitorFieldsV0
+	var browser *tfBrowserMonitorFieldsV0
 
 	switch mType := api.Type; mType {
 	case kbapi.Http:
@@ -428,6 +506,18 @@ func (v *tfModelV0) toModelV0(api *kbapi.SyntheticsMonitor) (*tfModelV0, error) 
 			tcp = v.TCP
 		}
 		tcp, err = tcp.toTfTCPMonitorFieldsV0(api)
+	case kbapi.Icmp:
+		icmp = &tfICMPMonitorFieldsV0{}
+		if v.ICMP != nil {
+			icmp = v.ICMP
+		}
+		icmp, err = icmp.toTfICMPMonitorFieldsV0(api)
+	case kbapi.Browser:
+		browser = &tfBrowserMonitorFieldsV0{}
+		if v.Browser != nil {
+			browser = v.Browser
+		}
+		browser, err = browser.toTfBrowserMonitorFieldsV0(api)
 	default:
 		err = fmt.Errorf("unsupported monitor type: %s", mType)
 	}
@@ -464,6 +554,8 @@ func (v *tfModelV0) toModelV0(api *kbapi.SyntheticsMonitor) (*tfModelV0, error) 
 		Params:           params,
 		HTTP:             http,
 		TCP:              tcp,
+		ICMP:             icmp,
+		Browser:          browser,
 		RetestOnFailure:  v.RetestOnFailure,
 	}, nil
 }
@@ -485,6 +577,47 @@ func (v *tfTCPMonitorFieldsV0) toTfTCPMonitorFieldsV0(api *kbapi.SyntheticsMonit
 		CheckReceive:          checkReceive,
 		ProxyURL:              types.StringValue(api.ProxyUrl),
 		ProxyUseLocalResolver: types.BoolPointerValue(api.ProxyUseLocalResolver),
+	}, nil
+}
+
+func (v *tfICMPMonitorFieldsV0) toTfICMPMonitorFieldsV0(api *kbapi.SyntheticsMonitor) (*tfICMPMonitorFieldsV0, error) {
+	wait, err := stringToInt64(string(api.Wait))
+	if err != nil {
+		return nil, err
+	}
+	return &tfICMPMonitorFieldsV0{
+		Host: types.StringValue(api.Host),
+		Wait: types.Int64Value(wait),
+	}, nil
+}
+
+func (v *tfBrowserMonitorFieldsV0) toTfBrowserMonitorFieldsV0(api *kbapi.SyntheticsMonitor) (*tfBrowserMonitorFieldsV0, error) {
+
+	var err error
+	playwrightOptions := v.PlaywrightOptions
+	if api.PlaywrightOptions != nil {
+		playwrightOptions, err = toNormalizedValue(api.PlaywrightOptions)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	syntheticsArgs := v.SyntheticsArgs
+	if api.SyntheticsArgs != nil {
+		syntheticsArgs = StringSliceValue(api.SyntheticsArgs)
+	}
+
+	inlineScript := v.InlineScript
+	if api.InlineScript != "" {
+		inlineScript = types.StringValue(api.InlineScript)
+	}
+
+	return &tfBrowserMonitorFieldsV0{
+		InlineScript:      inlineScript,
+		Screenshots:       types.StringValue(api.Screenshots),
+		SyntheticsArgs:    syntheticsArgs,
+		IgnoreHttpsErrors: types.BoolPointerValue(api.IgnoreHttpsErrors),
+		PlaywrightOptions: playwrightOptions,
 	}, nil
 }
 
@@ -572,6 +705,10 @@ func (v *tfModelV0) toMonitorFields() (kbapi.MonitorFields, diag.Diagnostics) {
 		return v.toHttpMonitorFields()
 	} else if v.TCP != nil {
 		return v.toTCPMonitorFields(), dg
+	} else if v.ICMP != nil {
+		return v.toICMPMonitorFields(), dg
+	} else if v.Browser != nil {
+		return v.toBrowserMonitorFields()
 	}
 
 	dg.AddError("Unsupported monitor type config", "one of http,tcp monitor fields is required")
@@ -606,6 +743,14 @@ func (v *tfModelV0) toSyntheticsMonitorConfig() (*kbapi.SyntheticsMonitorConfig,
 	}, diag.Diagnostics{} //dg
 }
 
+func tfInt64ToString(v types.Int64) string {
+	res := ""
+	if !v.IsUnknown() && !v.IsNull() { // handle omitempty case
+		return strconv.FormatInt(v.ValueInt64(), 10)
+	}
+	return res
+}
+
 func (v *tfModelV0) toHttpMonitorFields() (kbapi.MonitorFields, diag.Diagnostics) {
 	proxyHeaders, dg := toJsonObject(v.HTTP.ProxyHeader)
 	if dg.HasError() {
@@ -619,11 +764,7 @@ func (v *tfModelV0) toHttpMonitorFields() (kbapi.MonitorFields, diag.Diagnostics
 	if dg.HasError() {
 		return nil, dg
 	}
-	maxRedirects := ""
-	if !v.HTTP.MaxRedirects.IsUnknown() && !v.HTTP.MaxRedirects.IsNull() { // handle omitempty case
-		maxRedirects = strconv.FormatInt(v.HTTP.MaxRedirects.ValueInt64(), 10)
-
-	}
+	maxRedirects := tfInt64ToString(v.HTTP.MaxRedirects)
 	return kbapi.HTTPMonitorFields{
 		Url:                   v.HTTP.URL.ValueString(),
 		SslVerificationMode:   v.HTTP.SslVerificationMode.ValueString(),
@@ -651,6 +792,27 @@ func (v *tfModelV0) toTCPMonitorFields() kbapi.MonitorFields {
 		ProxyUrl:              v.TCP.ProxyURL.ValueString(),
 		ProxyUseLocalResolver: v.TCP.ProxyUseLocalResolver.ValueBoolPointer(),
 	}
+}
+
+func (v *tfModelV0) toICMPMonitorFields() kbapi.MonitorFields {
+	return kbapi.ICMPMonitorFields{
+		Host: v.ICMP.Host.ValueString(),
+		Wait: tfInt64ToString(v.ICMP.Wait),
+	}
+}
+
+func (v *tfModelV0) toBrowserMonitorFields() (kbapi.MonitorFields, diag.Diagnostics) {
+	playwrightOptions, dg := toJsonObject(v.Browser.PlaywrightOptions)
+	if dg.HasError() {
+		return nil, dg
+	}
+	return kbapi.BrowserMonitorFields{
+		InlineScript:      v.Browser.InlineScript.ValueString(),
+		Screenshots:       kbapi.ScreenshotOption(v.Browser.Screenshots.ValueString()),
+		SyntheticsArgs:    ValueStringSlice(v.Browser.SyntheticsArgs),
+		IgnoreHttpsErrors: v.Browser.IgnoreHttpsErrors.ValueBoolPointer(),
+		PlaywrightOptions: playwrightOptions,
+	}, diag.Diagnostics{} //dg
 }
 
 func Map[T, U any](ts []T, f func(T) U) []U {